Fix warning about comparing QtMsgType variable.

+++ Software/prismatic/LogWriter.cpp
@@ -63,7 +63,7 @@ void LogWriter::messageHandler(QtMsgType type, const QMessageLogContext &ctx, co
     static const LogWriter::Level s_msgType2Loglevel[] = {
         LogWriter::Debug, LogWriter::Warn, LogWriter::Critical, LogWriter::Fatal
     };
-    Q_ASSERT(type >= 0 && type < sizeof(s_msgType2Loglevel)/sizeof(s_msgType2Loglevel[0]));
+    Q_ASSERT(type >= 0 && static_cast<size_t>(type) < sizeof(s_msgType2Loglevel)/sizeof(s_msgType2Loglevel[0]));
     Q_UNUSED(ctx);
     if (g_logWriter)
         g_logWriter->writeMessage(msg, s_msgType2Loglevel[type]);
